services/health-api-service/requirements.txt
aio-pika==9.5.7
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiormq==6.9.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
attrs==25.3.0
avro==1.12.0
bcrypt==4.3.0
boto3==1.39.11
botocore==1.39.11
build==1.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
Deprecated==1.2.18
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.118.0
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
frozenlist==1.7.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
limits==5.6.0
makefun==1.16.0
multidict==6.6.4
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
pip-review==1.3.0
pip-tools==7.5.1
pluggy==1.6.0
prometheus_client==0.23.1
propcache==0.3.2
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-dotenv==0.5.2
pytest-env==1.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
structlog==25.4.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
